// src/ledger/ledger.did.d.ts
export interface Account {
  'owner' : Principal,
  'subaccount' : [] | [Array<number>],
}
export interface AccountBalanceArgs { 'account' : string }
export interface ArchiveInfo { 'canister_id' : Principal }
export interface ArchiveOptions {
  'num_blocks_to_archive' : bigint,
  'max_transactions_per_response' : [] | [bigint],
  'trigger_threshold' : bigint,
  'max_message_size_bytes' : [] | [bigint],
  'cycles_for_archive_creation' : [] | [bigint],
  'node_max_memory_size_bytes' : [] | [bigint],
  'controller_id' : Principal,
}
export interface ArchivedBlocksRange {
  'callback' : [Principal, string],
  'start' : bigint,
  'length' : bigint,
}
export interface Archives { 'archives' : Array<ArchiveInfo> }
export interface BinaryAccountBalanceArgs { 'account' : Array<number> }
export interface BlockRange { 'blocks' : Array<CandidBlock> }
export interface CandidBlock {
  'transaction' : CandidTransaction,
  'timestamp' : TimeStamp,
  'parent_hash' : [] | [Array<number>],
}
// ...
export interface Decimals { 'decimals' : number }
export interface Duration { 'secs' : bigint, 'nanos' : number }
export interface GetBlocksArgs { 'start' : bigint, 'length' : bigint }
// ...
export type MetadataValue = { 'Int' : bigint } |
  { 'Nat' : bigint } |
  { 'Blob' : Array<number> } |
  { 'Text' : string };
export interface Name { 'name' : string }
export interface QueryBlocksResponse {
  'certificate' : [] | [Array<number>],
  'blocks' : Array<CandidBlock>,
  'chain_length' : bigint,
  'first_block_index' : bigint,
  'archived_blocks' : Array<ArchivedBlocksRange>,
}
export type Result = { 'Ok' : bigint } |
  { 'Err' : TransferError };
export type Result_1 = { 'Ok' : bigint } |
  { 'Err' : TransferError_1 };
export interface SendArgs {
  'to' : string,
  'fee' : Tokens,
  'memo' : bigint,
  'from_subaccount' : [] | [Array<number>],
  'created_at_time' : [] | [TimeStamp],
  'amount' : Tokens,
}
export interface StandardRecord { 'url' : string, 'name' : string }
export interface Symbol { 'symbol' : string }
export interface TimeStamp { 'timestamp_nanos' : bigint }
export interface Tokens { 'e8s' : bigint }
export interface TransferArg {
  'to' : Account,
  'fee' : [] | [bigint],
  'memo' : [] | [Array<number>],
  'from_subaccount' : [] | [Array<number>],
  'created_at_time' : [] | [bigint],
  'amount' : bigint,
}
export interface TransferArgs {
  'to' : Array<number>,
  'fee' : Tokens,
  'memo' : bigint,
  'from_subaccount' : [] | [Array<number>],
  'created_at_time' : [] | [TimeStamp],
  'amount' : Tokens,
}
export type TransferError = {
    'GenericError' : { 'message' : string, 'error_code' : bigint }
  } |
  { 'TemporarilyUnavailable' : null } |
  { 'BadBurn' : { 'min_burn_amount' : bigint } } |
  { 'Duplicate' : { 'duplicate_of' : bigint } } |
  { 'BadFee' : { 'expected_fee' : bigint } } |
  { 'CreatedInFuture' : { 'ledger_time' : bigint } } |
  { 'TooOld' : null } |
  { 'InsufficientFunds' : { 'balance' : bigint } };
export type TransferError_1 = {
    'TxTooOld' : { 'allowed_window_nanos' : bigint }
  } |
  { 'BadFee' : { 'expected_fee' : Tokens } } |
  { 'TxDuplicate' : { 'duplicate_of' : bigint } } |
  { 'TxCreatedInFuture' : null } |
  { 'InsufficientFunds' : { 'balance' : Tokens } };
export interface TransferFee { 'transfer_fee' : Tokens }
export interface _SERVICE {
  'account_balance' : (arg_0: BinaryAccountBalanceArgs) => Promise<Tokens>,
  'account_balance_dfx' : (arg_0: AccountBalanceArgs) => Promise<Tokens>,
  'archives' : () => Promise<Archives>,
  'decimals' : () => Promise<Decimals>,
  'icrc1_balance_of' : (arg_0: Account) => Promise<bigint>,
  'icrc1_decimals' : () => Promise<number>,
  'icrc1_fee' : () => Promise<bigint>,
  'icrc1_metadata' : () => Promise<Array<[string, MetadataValue]>>,
  'icrc1_minting_account' : () => Promise<[] | [Account]>,
  'icrc1_name' : () => Promise<string>,
  'icrc1_supported_standards' : () => Promise<Array<StandardRecord>>,
  'icrc1_symbol' : () => Promise<string>,
  'icrc1_total_supply' : () => Promise<bigint>,
  'icrc1_transfer' : (arg_0: TransferArg) => Promise<Result>,
  'name' : () => Promise<Name>,
  'query_blocks' : (arg_0: GetBlocksArgs) => Promise<QueryBlocksResponse>,
  'send_dfx' : (arg_0: SendArgs) => Promise<bigint>,
  'symbol' : () => Promise<Symbol>,
  'transfer' : (arg_0: TransferArgs) => Promise<Result_1>,
  'transfer_fee' : (arg_0: {}) => Promise<TransferFee>,
}